Apply deprecation check to more API elements
